+++ tests/test_auth.py
@@ -7,7 +7,7 @@ from fractal_client.authclient import AuthenticationError
 from fractal_client.authclient import AuthToken
 
 
-async def test_auth_registered(client, register_user):
+def test_auth_registered(client, register_user):
     """
     GIVEN an existing user
     WHEN fetching a token
@@ -18,11 +18,11 @@ async def test_auth_registered(client, register_user):
         username=environ.get("FRACTAL_USER"),
         password=environ.get("FRACTAL_PASSWORD"),
     )
-    token = await auth()
+    token = auth()
     assert token
 
 
-async def test_auth_fail(client):
+def test_auth_fail(client):
     """
     GIVEN no user registered
     WHEN fetching a token
@@ -30,5 +30,5 @@ async def test_auth_fail(client):
     """
     with pytest.raises(AuthenticationError) as err:
         auth = AuthToken(client, username="foo", password="bar")
-        await auth()
+        auth()
     debug(err.value)