+++ tests/test_project.py
@@ -2,76 +2,76 @@ import pytest
 from devtools import debug
 
 
-async def test_project_create(register_user, invoke):
+def test_project_create(register_user, invoke):
     PROJECT_NAME = "project_name"
-    res = await invoke(f"project new {PROJECT_NAME}")
+    res = invoke(f"project new {PROJECT_NAME}")
     debug(res)
     assert res.data["name"] == PROJECT_NAME
 
 
-async def test_project_delete(register_user, invoke):
+def test_project_delete(register_user, invoke):
 
     # Create project
-    res = await invoke("project new MyProj1")
+    res = invoke("project new MyProj1")
     res.show()
     project_id_1 = res.data["id"]
 
     # Show project
-    res = await invoke(f"project show {project_id_1}")
+    res = invoke(f"project show {project_id_1}")
     res.show()
 
     # Delete project
-    res = await invoke(f"project delete {project_id_1}")
+    res = invoke(f"project delete {project_id_1}")
     assert res.retcode == 0
 
     # Try to show deleted project, and fail
     with pytest.raises(SystemExit):
-        res = await invoke(f"project show {project_id_1}")
+        res = invoke(f"project show {project_id_1}")
 
 
-async def test_project_create_batch(register_user, invoke):
-    res = await invoke("--batch project new MyProj1")
+def test_project_create_batch(register_user, invoke):
+    res = invoke("--batch project new MyProj1")
     debug(res)
     debug(res.data)
     project_id = int(res.data)
     assert project_id == 1
 
 
-async def test_project_list(register_user, invoke):
-    res = await invoke("project list")
+def test_project_list(register_user, invoke):
+    res = invoke("project list")
     debug(res)
     debug(vars(res.data))
     assert len(res.data.rows) == 0
 
     res.show()
 
-    res = await invoke("--batch project new proj0")
+    res = invoke("--batch project new proj0")
     project0_id = res.data
-    res = await invoke(f"--batch project add-dataset {project0_id} NAME")
-    res = await invoke("--batch project new proj1")
+    res = invoke(f"--batch project add-dataset {project0_id} NAME")
+    res = invoke("--batch project new proj1")
 
-    res = await invoke("project list")
+    res = invoke("project list")
     debug(res)
     debug(vars(res.data))
     res.show()
     assert len(res.data.rows) == 2
 
 
-async def test_add_dataset(register_user, invoke):
+def test_add_dataset(register_user, invoke):
     DATASET_NAME = "new_ds_name"
 
-    res = await invoke("--batch project new proj0")
+    res = invoke("--batch project new proj0")
     assert res.retcode == 0
     debug(res.data)
     project_id = int(res.data)
 
-    res = await invoke(f"project add-dataset {project_id} {DATASET_NAME}")
+    res = invoke(f"project add-dataset {project_id} {DATASET_NAME}")
     assert res.retcode == 0
     res.show()
     assert res.data["name"] == DATASET_NAME
     assert not res.data["read_only"]
 
-    res = await invoke(
+    res = invoke(
         f"project add-dataset {project_id} new_{DATASET_NAME} --make-read-only"
     )
     assert res.retcode == 0
@@ -82,7 +82,7 @@ async def test_add_dataset(register_user, invoke):
 
 @pytest.mark.parametrize("new_name", ["new_name", None])
 @pytest.mark.parametrize("read_only", [True, False, None])
-async def test_edit_project(
+def test_edit_project(
     register_user,
     invoke,
     new_name,
@@ -90,7 +90,7 @@ async def test_edit_project(
     tmp_path,
 ):
     name = "name"
-    res = await invoke(f"project new {name}")
+    res = invoke(f"project new {name}")
     project = res.data
     project_id = project["id"]
 
@@ -102,7 +102,7 @@ async def test_edit_project(
     elif read_only is False:
         cmd += " --remove-read-only"
 
-    res = await invoke(cmd)
+    res = invoke(cmd)
     debug(res)
 
     if (not new_name) and (read_only is None):