+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # 1.4.3 (unreleased)
 
+* Make `fractal-client` a fully synchronous client, by removing all `async`/`await` (\#592).
 * Improve handling of `AuthenticationError` and `ConnectionError`, and handle uncaught `Exception` (\#587).
 * Testing:
     * Adapt `job_factory` and tests to the presence of new `timestamp_created` attributes in `fractal-server` (\#589).